+++ src/grid/Component.js
@@ -158,8 +158,7 @@ const GridComponent = createReactClass({
   componentDidMount: function () {
     this._isMounted = true;
     if (this.props.model) {
-      this.props.model.on('create', this._onRecordCreated);
-      this.props.model.on('createRecords', this._onRecordsCreated);
+      this.props.model.on('create', this._onRecordsCreated);
       this.props.model.on('update', this._setData);
       this.props.model.on('delete', this.updateTable);
     }
@@ -168,8 +167,7 @@ const GridComponent = createReactClass({
   componentWillUnmount: function () {
     this._isMounted = false;
     if (this.props.model) {
-      this.props.model.off('create', this._onRecordCreated);
-      this.props.model.off('createRecords', this._onRecordsCreated);
+      this.props.model.off('create', this._onRecordsCreated);
       this.props.model.off('update', this._setData);
       this.props.model.off('delete', this.updateTable);
     }
@@ -215,13 +213,11 @@ const GridComponent = createReactClass({
         if (reset & RESET_MODEL) {
           this.state.data = null;
           if (oldProps.model) {
-            oldProps.model.off('create', this._onRecordCreated);
-            oldProps.model.off('createRecords', this._onRecordsCreated);
+            oldProps.model.off('create', this._onRecordsCreated);
             oldProps.model.off('update', this._setData);
           }
           if (this.props.model) {
-            this.props.model.on('create', this._onRecordCreated);
-            oldProps.model.on('createRecords', this._onRecordsCreated);
+            this.props.model.on('create', this._onRecordsCreated);
             this.props.model.on('update', this._setData);
           }
           this._setPage(0);